# docs/api/phase-4-openapi.yaml
openapi: 3.0.3
info:
  title: TeamFlow API
  version: 0.1.0
  description: Phase 4 project, task, label, and comment contract.
servers:
  - url: http://localhost:4000
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /organizations/{organizationId}/tasks:
    get:
      summary: List organization tasks
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - { name: projectId, in: query, schema: { type: string, format: uuid } }
        - {
            name: assigneeId,
            in: query,
            schema: { type: string, format: uuid },
          }
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [todo, in_progress, done] },
          }
        - { name: search, in: query, schema: { type: string } }
      responses: { "200": { $ref: "#/components/responses/PaginatedResponse" } }
    post:
      summary: Create a task
      description: Requires owner, admin, or member membership.
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { $ref: "#/components/requestBodies/TaskInput" }
      responses: { "201": { $ref: "#/components/responses/TaskResponse" } }
  /organizations/{organizationId}/tasks/{taskId}:
    get:
      summary: Get a task
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/TaskId"
      responses: { "200": { $ref: "#/components/responses/TaskResponse" } }
    patch:
      summary: Update a task
      description: Requires owner, admin, or member membership.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/TaskId"
      requestBody: { $ref: "#/components/requestBodies/TaskUpdate" }
      responses: { "200": { $ref: "#/components/responses/TaskResponse" } }
    delete:
      summary: Delete a task
      description: Requires owner, admin, or member membership.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/TaskId"
      responses: { "200": { $ref: "#/components/responses/DeletedResponse" } }
  /organizations/{organizationId}/labels:
    get:
      summary: List organization labels
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses: { "200": { $ref: "#/components/responses/PaginatedResponse" } }
    post:
      summary: Create a label
      description: Requires owner or admin membership.
      parameters: [{ $ref: "#/components/parameters/OrganizationId" }]
      requestBody: { $ref: "#/components/requestBodies/LabelInput" }
      responses: { "201": { $ref: "#/components/responses/LabelResponse" } }
  /organizations/{organizationId}/tasks/{taskId}/comments:
    get:
      summary: List task comments
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses: { "200": { $ref: "#/components/responses/PaginatedResponse" } }
    post:
      summary: Add a task comment
      description: Requires owner, admin, or member membership.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/TaskId"
      requestBody: { $ref: "#/components/requestBodies/CommentInput" }
      responses: { "201": { $ref: "#/components/responses/CommentResponse" } }
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: teamflow_access_token }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    OrganizationId:
      {
        name: organizationId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    TaskId:
      {
        name: taskId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
    Page:
      {
        name: page,
        in: query,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    PageSize:
      {
        name: pageSize,
        in: query,
        schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
      }
  requestBodies:
...
    TaskInput:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/TaskInput" } },
        }
    TaskUpdate:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/TaskUpdate" } },
        }
    LabelInput:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/LabelInput" } },
        }
    CommentInput:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/CommentInput" } },
        }
  responses:
...
    TaskResponse:
      {
        description: Task resource,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [task],
                    properties: { task: { $ref: "#/components/schemas/Task" } },
                  },
              },
          },
      }
    LabelResponse:
      {
        description: Label resource,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [label],
                    properties:
                      { label: { $ref: "#/components/schemas/Label" } },
                  },
              },
          },
      }
    CommentResponse:
      {
        description: Comment resource,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [comment],
                    properties:
                      { comment: { $ref: "#/components/schemas/Comment" } },
                  },
              },
          },
      }
...
    PaginatedResponse:
      {
        description: Paginated list response,
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [items, pagination],
                    properties:
                      {
                        items: { type: array, items: {} },
                        pagination: { $ref: "#/components/schemas/Pagination" },
                      },
                  },
              },
          },
      }
    DeletedResponse:
      {
        description: Deletion response,
        content:
          {
            application/json:
              {
                schema:
                  { type: object, properties: { deleted: { type: boolean } } },
              },
          },
      }
  schemas:
...
    TaskInput:
      type: object
      required: [projectId, title]
      properties:
        {
          projectId: { type: string, format: uuid },
          title: { type: string },
          description: { type: string, nullable: true },
          status: { type: string, enum: [todo, in_progress, done] },
          dueDate: { type: string, format: date-time, nullable: true },
          assigneeId: { type: string, format: uuid, nullable: true },
        }
    TaskUpdate: { $ref: "#/components/schemas/TaskInput" }
    LabelInput:
      type: object
      required: [name, color]
      properties:
        {
          name: { type: string },
          color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" },
        }
    CommentInput:
      type: object
      required: [body]
      properties: { body: { type: string, minLength: 1, maxLength: 5000 } }
...
    Task:
      {
        type: object,
        properties:
          {
            id: { type: string, format: uuid },
            organizationId: { type: string, format: uuid },
            projectId: { type: string, format: uuid },
            title: { type: string },
            status: { type: string },
            dueDate: { type: string, format: date-time, nullable: true },
            assigneeId: { type: string, format: uuid, nullable: true },
          },
      }
    Label:
      {
        type: object,
        properties:
          {
            id: { type: string, format: uuid },
            organizationId: { type: string, format: uuid },
            name: { type: string },
            color: { type: string },
          },
      }
    Comment:
      {
        type: object,
        properties:
          {
            id: { type: string, format: uuid },
            organizationId: { type: string, format: uuid },
            taskId: { type: string, format: uuid },
            authorId: { type: string, format: uuid },
            body: { type: string },
          },
      }
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        {
          page: { type: integer },
          pageSize: { type: integer },
          total: { type: integer },
        }